cleanup(ac-ui): remove dead code, unused dep, and unreachable layout
- Remove crossbeam-channel dependency (never imported)
- Remove LayoutMode::Overlay variant and all handling (unreachable via
  L-cycle since it was excluded; removes ~30 lines across 5 files)
- Remove DisplayFrame::peak_hold and all peak decay computation in
  ChannelStore (data was computed every frame but never rendered)
- Remove DisplayConfig::peak_hold and P keybinding (toggled dead feature)
- Remove SpectrumRenderer::skip_fill/set_skip_fill (only served overlay)
- Remove overlay fill-alpha/line-width computation in redraw hot path
- Remove redundant cells.clone() in redraw
- Remove debug-artifact test reading /tmp/claude/transfer_body.json
- Remove incorrect #[allow(dead_code)] on DataSource::status() (live method)
- Fix clippy: unnecessary u32 casts in synthetic.rs, unwrap-after-is_none
  in overlay transfer section

+++ ac-rs/crates/ac-ui/src/data/synthetic.rs
@@ -178,7 +178,7 @@ fn synth_transfer(freqs: &[f32], meas: u32, refc: u32, t: f32) -> TransferFrame 
     let f_lp = 15000.0;
 
     let mut rng = XorShift::new(
-        0xA17E ^ (meas as u32) ^ ((refc as u32) << 8) ^ ((t * 10.0) as u32),
+        0xA17E ^ meas ^ (refc << 8) ^ ((t * 10.0) as u32),
     );
 
     for &f in freqs {
